# syllabus/python-full-stack/assests/module-16-spec-kit/code-openapi-spec-example.yaml
...
openapi: 3.0.3
info:
  title: TechPath Student Management API
  version: 1.0.0
  description: |
    REST API for managing students, batches, and attendance at TechPath Institute, Bhopal.

    **Authentication:** All protected endpoints require a Bearer token in the Authorization header.

    **Response Format:** All responses follow the standard format:
    ```json
    {
      "success": true,
      "data": { ... },
      "timestamp": "2026-07-25T10:30:00Z"
    }
    ```
  contact:
    name: TechPath Institute
    email: support@techpath.biz
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://staging.api.techpath.biz/api/v1
    description: Staging server (for testing)
  - url: https://api.techpath.biz/api/v1
    description: Production server

# ---------------------------------------------------------------------------
# Authentication Schemes
# ---------------------------------------------------------------------------
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token or JWT from /auth/login

  # -------------------------------------------------------------------------
  # Reusable Schemas (Data Models)
  # -------------------------------------------------------------------------
  schemas:
...
    Student:
      type: object
      required: [name, email]
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 100
          example: "Rahul Sharma"
        email:
          type: string
          format: email
          example: "rahul.sharma@example.com"
        phone:
          type: string
          pattern: "^[6-9]\\d{9}$"
          example: "9876543210"
          description: "10-digit Indian mobile number"
        batch_id:
          type: integer
          example: 1
        city:
          type: string
          example: "Bhopal"
        is_active:
          type: boolean
          default: true
        enrolled_at:
          type: string
          format: date-time
          readOnly: true
    StudentCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          maxLength: 100
          example: "Priya Patel"
        email:
          type: string
          format: email
          example: "priya.patel@example.com"
        phone:
          type: string
          example: "8765432109"
        batch_id:
          type: integer
          example: 1
        city:
          type: string
          example: "Pune"
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        phone:
          type: string
        city:
          type: string
        is_active:
          type: boolean

    # --- Batch ---
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Student not found"
        timestamp:
          type: string
          format: date-time
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Attendance marked successfully"
        timestamp:
          type: string
          format: date-time

  # -------------------------------------------------------------------------
  # Reusable Parameters
  # -------------------------------------------------------------------------
  parameters:
    skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of records to skip (for pagination)
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum records to return

# ---------------------------------------------------------------------------
# API Endpoints (Paths)
# ---------------------------------------------------------------------------
paths:
...
  /students:
    get:
      summary: List all students
      description: Returns a paginated list of students. Supports filtering by batch and city.
      tags: [Students]
      parameters:
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/limit"
        - name: batch_id
          in: query
          schema:
            type: integer
          description: Filter by batch ID
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city name
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        "200":
          description: List of students
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total number of matching students
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"

    post:
      summary: Create a new student
      tags: [Students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentCreate"
            example:
              name: "Ananya Gupta"
              email: "ananya.gupta@example.com"
              phone: "7654321098"
              batch_id: 1
              city: "Delhi"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Student"
        "409":
          description: Student with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{student_id}:
    get:
      summary: Get a student by ID
      tags: [Students]
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Update a student
      tags: [Students]
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Student"
        "404":
          description: Student not found

    delete:
      summary: Soft-delete a student (sets is_active=false)
      tags: [Students]
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- Attendance ---
...